+++ core/src/main/kotlin/com/retoday/core/domain/history/entity/Page.kt
@@ -12,9 +12,8 @@ class Page(
     @Id
     @Tsid
     val id: Long? = null,
-    @Column(nullable = false)
     val websiteId: Long,
-    @Column(nullable = false, length = 2048, unique = true)
+    @Column(length = 2048, unique = true)
     val url: String,
     @Column(length = 500)
     var title: String? = null,
@@ -22,9 +21,7 @@ class Page(
     var description: String? = null,
     @Column(length = 500)
     var thumbnailUrl: String? = null,
-    @Column(nullable = false)
     var totalVisitCount: Long = 0,
-    @Column(nullable = false)
     val firstVisitedAt: Instant,
     var lastVisitedAt: Instant? = null
 ) : BaseEntity() {
