+++ core/src/test/kotlin/com/retoday/core/domain/history/repository/CustomWebsiteRepositoryTest.kt
@@ -1,8 +1,7 @@
-package com.retoday.core.domain.website.repository
+package com.retoday.core.domain.history.repository
 
 import com.retoday.core.common.RepositoryTest
 import com.retoday.core.domain.history.entity.Website
-import com.retoday.core.domain.history.repository.CustomWebsiteRepositoryImpl
 import com.retoday.core.domain.user.entity.User
 import com.retoday.core.domain.user.entity.UserExcludedWebsite
 import com.retoday.core.fixture.createUser
